requirements.txt
cachelib==0.12.0
mlflow>=2.11.1
oauthlib==3.2.2
python-dotenv==1.0.0
requests==2.31.0
sqlalchemy==2.0.28
flask>=2.3.2
Flask-Session>=0.7.0
#flask_sqlalchemy>=3.1.1
#flask-talisman>=1.1.0
#flask-healthz>=0.0.3
#flask-wtf>=1.2.1
#waitress==2.1.2
